Fix some issues with extension messaging for transient pages. - Calling sendResponse within an onMessage listener no longer throws an error. - Sending a message to a transient page that is only partially loaded no longer fails.
BUG=120531
TEST=no

Review URL: https://chromiumcodereview.appspot.com/9968068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130417 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d79e3ab0a375657890bbcdf13aa7f22014ab5f22

+++ chrome/renderer/resources/extensions/miscellaneous_bindings.js
@@ -162,7 +162,7 @@
             var rv = requestEvent.dispatch(request, sender, responseCallback);
             responseCallbackPreserved =
                 rv && rv.results && rv.results.indexOf(true) > -1;
-            if (!responseCallbackPreserved) {
+            if (!responseCallbackPreserved && port) {
               // If they didn't access the response callback, they're not
               // going to send a response, so clean up the port immediately.
               port.destroy_();
